requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
ecdsa==0.19.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
